requirements.txt
web3==6.11.1
python-dotenv==1.0.0
requests==2.31.0
pandas==2.1.3
numpy==1.26.2
fastapi==0.104.1
uvicorn==0.24.0
pytest==7.4.3
python-dateutil==2.8.2
aiohttp==3.9.1 